feat: core market-data REST endpoints — intraday, dividends, live, search, exchanges, IPOs (PR 3a/N) (#3)
**Phase 3a of the EODHD OSS client** — first slice of the
full-REST-coverage work. Completes the orphan models already sitting in
the tree and adds the core market-data endpoints.

### New endpoints (each `async` + sync wrapper)
- **Intraday historical** — `GetIntradayHistoricalStockPrices[Async]`
(`interval` 1m/5m/1h, Unix `from`/`to`). Wires the existing
`IntradayHistoricalStockPrice` model.
- **Historical dividends** — `GetHistoricalDividends[Async]` (the `div`
endpoint). Wires the existing `HistoricalDividend` model.
- **Live / delayed real-time** — `GetLivePrice[Async]` (single) and
`GetLivePrices[Async]` (multi-symbol via `s=`). New `LiveStockPrice`
model.
- **Search** — `Search[Async]` (limit/type/exchange/bonds_only). New
`SearchResult` model.
- **Exchanges list** — `GetExchangesList[Async]`. Uses the existing
`Exchange` model.
- **IPO calendar** — `GetUpcomingIpos[Async]`. New `Ipo` /
`UpcomingIpos` models.

### Two converters for EODHD wire quirks
- `SpaceSeparatedDateTimeConverter` — intraday `datetime` is
`"yyyy-MM-dd HH:mm:ss"` (not ISO-8601, which STJ rejects); also accepts
ISO, date-only, Unix-seconds, and `null`/`"NA"`.
- `NaTolerantDoubleConverter` — live quote numerics come back as the
literal `"NA"` outside trading hours → mapped to `null`.

### Tests
- **Unit (mock transport):** both converters, every new endpoint, and
URL/param assertions (e.g. `interval=1h`, `s=MSFT.US`, `limit`/`type`).
**20 pass.**
- **Integration (`SkippableFact`, gated on `EODHD_API_KEY`):** one per
new endpoint. **24 skip cleanly** without a key.

### Verified locally
- `dotnet build -c Release` → **0 warnings, 0 errors** (CS1591-as-error
→ every public member documented)
- Unit **20/20**, integration **24 skipped**

### Note on verification
Field mappings follow the EODHD REST documentation. The integration
suite is written to verify them against live data — **worth running once
with the production key before v1.0.0**, since these endpoints haven't
yet been exercised against live responses in CI.

### Remaining REST surface (subsequent sub-PRs)
Technical indicators, screener, news, sentiment (news + social), macro
indicators, economic events, bonds fundamentals, options, insider
transactions, historical market cap, and the earnings-trends calendar
(`calendar/trends` — deferred until its nested-array shape is verified
live). Then PR 4: Live/WebSocket API.

+++ src/EODHD.CSharpApiClient/DataModel/BulkFundamental/Highlights.cs
@@ -18,7 +18,7 @@ namespace EODHD.CSharpApiClient.DataModel.BulkFundamental
         /// <summary>
         /// Gets or sets the EBITDA (currency).
         /// </summary>
-        public long? EBITDA { get; set; }
+        public decimal? EBITDA { get; set; }
 
         /// <summary>
         /// Gets or sets the price-to-earnings ratio.
@@ -103,7 +103,7 @@ namespace EODHD.CSharpApiClient.DataModel.BulkFundamental
         /// <summary>
         /// Gets or sets the trailing-twelve-month revenue (currency).
         /// </summary>
-        public long? RevenueTTM { get; set; }
+        public decimal? RevenueTTM { get; set; }
 
         /// <summary>
         /// Gets or sets the trailing-twelve-month revenue per share (currency).
@@ -118,7 +118,7 @@ namespace EODHD.CSharpApiClient.DataModel.BulkFundamental
         /// <summary>
         /// Gets or sets the trailing-twelve-month gross profit (currency).
         /// </summary>
-        public long? GrossProfitTTM { get; set; }
+        public decimal? GrossProfitTTM { get; set; }
 
         /// <summary>
         /// Gets or sets the trailing-twelve-month diluted earnings per share (currency).
