+++ src/EODHD.CSharpApiClient/DataModel/BulkFundamental/Technicals.cs
@@ -39,12 +39,12 @@ namespace EODHD.CSharpApiClient.DataModel.BulkFundamental
         /// <summary>
         /// Gets or sets the number of shares short (count).
         /// </summary>
-        public long? SharesShort { get; set; }
+        public decimal? SharesShort { get; set; }
 
         /// <summary>
         /// Gets or sets the number of shares short in the prior month (count).
         /// </summary>
-        public long? SharesShortPriorMonth { get; set; }
+        public decimal? SharesShortPriorMonth { get; set; }
 
         /// <summary>
         /// Gets or sets the short ratio.
